+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - This Changelog.
 - Automated release process on Git Tag push.
 - `--version` command line option.
+- The exporter follows `include` directives in the IPsec config file and searches for connections in all referenced
+  files.
 
 ### Changed
 - The default listening port is now `9536` and registered as a [Prometheus Exporter Default port](https://github.com/prometheus/prometheus/wiki/Default-port-allocations).